clean unused code and fix submission final grade logic
We had a weight loop which was not being used anywhere so i deleted it.
Now we add the submission weight based on the problem and not the
submission itself. We do that by going though every final submissions
and adding the final grade and the weight to that submission, before
displaying it to the user.